movidos programas desperdicio para o modulo desperdicio
