+++ sme_terceirizadas/cardapio/models.py
@@ -1478,46 +1478,3 @@ class MotivoDRENaoValida(
     class Meta:
         verbose_name = 'Motivo de não validação da DRE'
         verbose_name_plural = 'Motivos de não validação da DRE'
-
-
-class ControleSobras(TemChaveExterna, CriadoEm):
-    escola = models.ForeignKey('escola.Escola', on_delete=models.DO_NOTHING)
-    tipo_alimentacao = models.ForeignKey(TipoAlimentacao, on_delete=models.DO_NOTHING)
-    tipo_alimento = models.ForeignKey('produto.TipoAlimento', on_delete=models.DO_NOTHING)
-    peso_alimento = models.DecimalField(max_digits=5, decimal_places=2)
-    tipo_recipiente = models.ForeignKey('produto.TipoRecipiente', on_delete=models.DO_NOTHING)
-    peso_recipiente = models.DecimalField(max_digits=5, decimal_places=2)
-    peso_sobra = models.DecimalField(max_digits=5, decimal_places=2)
-    usuario = models.ForeignKey('perfil.Usuario', default='', on_delete=models.DO_NOTHING)
-    data_hora_medicao = models.DateTimeField('Data e hora da medição', null=True, default=None)
-
-    class Meta:
-        verbose_name = 'Controle de Sobras'
-
-
-class ControleRestos(TemChaveExterna, CriadoEm):
-    escola = models.ForeignKey('escola.Escola', on_delete=models.DO_NOTHING)
-    tipo_alimentacao = models.ForeignKey(TipoAlimentacao, on_delete=models.DO_NOTHING)
-    peso_resto = models.DecimalField(max_digits=5, decimal_places=2)
-    cardapio = models.CharField('cardapio', max_length=250, blank=True)
-    resto_predominante = models.CharField('resto_predominante', max_length=250, blank=True)
-    usuario = models.ForeignKey('perfil.Usuario', default='', on_delete=models.DO_NOTHING)
-    data_hora_medicao = models.DateTimeField('Data e hora da medição', null=True, default=None)
-    quantidade_distribuida = models.DecimalField(max_digits=5, decimal_places=2)
-
-    @property
-    def imagens(self):
-        return self.imagemcontroleresto_set.all()
-
-    class Meta:
-        verbose_name = 'Controle de Restos'
-
-
-class ImagemControleResto(TemChaveExterna):
-    controle_resto = models.ForeignKey('ControleRestos', on_delete=models.CASCADE, blank=True)
-    arquivo = models.FileField()
-    nome = models.CharField(max_length=255, blank=True)
-
-    class Meta:
-        verbose_name = 'Imagem do Controle de Resto'
-        verbose_name_plural = 'Imagens do Controle de Resto'
