+++ sme_terceirizadas/cardapio/tasks.py
@@ -1,4 +1,3 @@
-import io
 import datetime
 import logging
 
@@ -15,14 +14,6 @@ from ..escola.models import (
     PeriodoEscolar,
     TipoUnidadeEscolar,
 )
-from sme_terceirizadas.dados_comuns.utils import (
-    atualiza_central_download,
-    atualiza_central_download_com_erro,
-    build_xlsx_generico,
-    gera_objeto_na_central_download
-)
-
-from .api.serializers.serializers import serialize_relatorio_controle_restos, serialize_relatorio_controle_sobras
 from .models import VinculoTipoAlimentacaoComPeriodoEscolarETipoUnidadeEscolar
 
 logger = logging.getLogger("sigpae.taskCardapio")
@@ -96,66 +87,3 @@ def ativa_desativa_vinculos_alimentacao_com_periodo_escolar_e_tipo_unidade_escol
             # deve ativar INTEGRAL, MANHA E TARDE PARA CEIS DA PROPERTY CRIADA
             periodos_escolares = PERIODOS_ESPECIAIS_CEI_DIRET
             bypass_ativa_vinculos(tipo_unidade, periodos_escolares)
-
-
-@shared_task(
-    retry_backoff=2,
-    retry_kwargs={'max_retries': 8},
-    time_limit=3000,
-    soft_time_limit=3000
-)
-def gera_xls_relatorio_controle_restos_async(user, nome_arquivo, data):
-    logger.info(f'x-x-x-x Iniciando a geração do arquivo {nome_arquivo} x-x-x-x')
-    obj_central_download = gera_objeto_na_central_download(user=user, identificador=nome_arquivo)
-    try:
-        output = io.BytesIO()
-
-        titulos_colunas = [
-            'DRE', 'Unidade Educacional', 'Data da Medição',
-            'Quantidade Distribuída', 'Peso do Resto (Kg)', 'Nº Refeições',
-            'Resto per Capita', '% Resto']
-        build_xlsx_generico(
-            output,
-            queryset_serializada=[serialize_relatorio_controle_restos(item) for item in data],
-            titulo='Relatório de Controle de Restos',
-            titulo_sheet='Relatório',
-            titulos_colunas=titulos_colunas,
-        )
-
-        atualiza_central_download(obj_central_download, nome_arquivo, output.read())
-    except Exception as e:
-        atualiza_central_download_com_erro(obj_central_download, str(e))
-
-    logger.info(f'x-x-x-x Finaliza a geração do arquivo {nome_arquivo} x-x-x-x')
-
-
-@shared_task(
-    retry_backoff=2,
-    retry_kwargs={'max_retries': 8},
-    time_limit=3000,
-    soft_time_limit=3000
-)
-def gera_xls_relatorio_controle_sobras_async(user, nome_arquivo, data):
-    logger.info(f'x-x-x-x Iniciando a geração do arquivo {nome_arquivo} x-x-x-x')
-    obj_central_download = gera_objeto_na_central_download(user=user, identificador=nome_arquivo)
-    try:
-        output = io.BytesIO()
-
-        titulos_colunas = [
-            'DRE', 'Unidade Educacional', 'Tipo de Alimentação', 'Tipo de Alimento',
-            'Data da Medição', 'Peso da Refeição Distribuída (Kg)', 'Peso da Sobra (Kg)',
-            'Total de Alunos (frequência)', 'Total Primeira Oferta', 'Total Segunda Oferta (Repetição)',
-            '% Sobra', 'Média por Aluno', 'Média por Refeição']
-        build_xlsx_generico(
-            output,
-            queryset_serializada=[serialize_relatorio_controle_sobras(item) for item in data],
-            titulo='Relatório de Controle de Sobras',
-            titulo_sheet='Relatório',
-            titulos_colunas=titulos_colunas,
-        )
-
-        atualiza_central_download(obj_central_download, nome_arquivo, output.read())
-    except Exception as e:
-        atualiza_central_download_com_erro(obj_central_download, str(e))
-
-    logger.info(f'x-x-x-x Finaliza a geração do arquivo {nome_arquivo} x-x-x-x')
