+++ sme_terceirizadas/desperdicio/constants.py
@@ -1,1 +1,11 @@
+M = "Manhã"
+T = "Tarde"
+I = "Integral"
+
+SOBRAS_PERIODO = ["M", "T", "I"]
+
+CR = "CLASSIFICAÇÃO DE RESTO"
+CS = "CLASSIFICAÇÃO DE SOBRA"
+AR = "ACEITABILIDADE DAS REFEIÇÕES"
+
 CLASSIFICACAO_TIPO = ["CR", "CS", "AR"]
