+++ sme_terceirizadas/desperdicio/urls.py
@@ -5,6 +5,16 @@ from .api import viewsets
 
 router = routers.DefaultRouter()
 router.register(
+    'controle-sobras',
+    viewsets.ControleSobrasViewSet,
+    'controle-sobras'
+)
+router.register(
+    'controle-restos',
+    viewsets.ControleRestosViewSet,
+    'controle-restos'
+)
+router.register(
     "parametros-classificacao", 
     viewsets.ClassificacaoViewSet, 
     "Parâmetros de Classificação"
